# apps/backend/entities/CheckinRecord.ts
import {
  Entity,
  PrimaryGeneratedColumn,
  Column,
  ManyToOne,
  CreateDateColumn,
} from 'typeorm';
import { User } from './User';
import { Habit } from './Habit';

@Entity()
export class CheckinRecord {
  @PrimaryGeneratedColumn()
  id!: number;

  @Column({ type: 'date' })
  date!: string;

  @Column({ nullable: true })
  note?: string;

  @Column({ nullable: true })
  imageUrl?: string;

  @CreateDateColumn()
  createdAt!: Date;

  @ManyToOne(() => User)
  user!: User;

  @ManyToOne(() => Habit, habit => habit.records)
  habit!: Habit;
}