# apps/backend/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Habit": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"icon": {
						"type": "string"
					},
					"frequency": {
						"type": "string"
					},
					"remindTime": {
						"type": "string"
					},
					"targetDays": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"records": {
						"items": {
							"$ref": "#/components/schemas/CheckinRecord"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"icon",
					"frequency",
					"createdAt",
					"user",
					"records"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"openid": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"nickname": {
						"type": "string"
					},
					"habits": {
						"items": {
							"$ref": "#/components/schemas/Habit"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"nickname",
					"habits"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CheckinRecord": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string"
					},
					"note": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"habit": {
						"$ref": "#/components/schemas/Habit"
					}
				},
				"required": [
					"id",
					"date",
					"createdAt",
					"user",
					"habit"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/habits/{userId}": {
			"get": {
				"operationId": "GetHabits",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Habit"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Habit"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/habits": {
			"post": {
				"operationId": "CreateHabit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Habit"
								}
							}
						}
					}
				},
				"tags": [
					"Habit"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Habit"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Habit"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Habit"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}